create rgister user functionality

+++ admin/includes/view_all_post.php
@@ -103,7 +103,7 @@
                 echo "<td>$post_date</td>";
                 echo "<td><a href='../post.php?p_id={$post_id}'>View Post</a></td>";
                 echo "<td><a href='posts.php?source=edit_post&p_id={$post_id}'>Edit</a></td>";
-                echo "<td><a href='posts.php?delete={$post_id}'>Delete</a></td>";
+                echo "<td><a onClick=\"javascript: return confirm('are you sure you want to delete this post?'); \" href='posts.php?delete={$post_id}'>Delete</a></td>";
                 echo "</tr>";
             }
             ?>
@@ -117,6 +117,6 @@ if (isset($_GET['delete'])) {
     $the_post_id = $_GET['delete'];
     $query = "DELETE FROM posts WHERE post_id LIKE {$the_post_id} ";
     $delete_query = mysqli_query($conection, $query);
-    header("Location: view_all_post.php");
+    header("Location: posts.php");
 }
 ?>
